DX12: dynamic memory allocator redesign

+++ DX12/D3D12Research/D3D12/Graphics/CommandContext.cpp
@@ -122,7 +122,7 @@ void CommandContext::FlushResourceBarriers()
 	}
 }
 
-void CommandContext::InitializeBuffer(GraphicsBuffer* pResource, const void* pData, uint32_t dataSize, uint32_t offset)
+void CommandContext::InitializeBuffer(GraphicsBuffer* pResource, const void* pData, uint64_t dataSize, uint32_t offset)
 {
 	DynamicAllocation allocation = m_DynamicAllocator->Allocate(dataSize);
 	memcpy(allocation.pMappedMemory, pData, dataSize);
